requirements.txt
# Raspberry Pi Voice Assistant Dependencies

# LLM Inference
llama-cpp-python==0.2.20

# Speech-to-Text
vosk==0.3.45

# Audio Processing
pyaudio==0.2.14
numpy==1.24.3

# Text-to-Speech
pyttsx3==2.90

# Voice Activity Detection (optional but recommended)
webrtcvad==2.0.10

# Configuration
pyyaml==6.0.1

# HTTP Requests (for laptop client and Serper API)
requests>=2.31.0

# Utilities
colorlog==6.7.0

